pyproject.toml
[tool.poetry]
name = "astrodown"
version = "0.1.6"
description = ""
authors = ["Qiushi Yan <qiushi.yann@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pyodide-http = "^0.2.0"
typer = {extras = ["all"], version = "^0.7.0"}
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
pytest = "^7.2"
bpython = "^0.24"
pandas = "^1.5.3"
pytest-pyodide = "^0.50.0"

[tool.ruff]
line-length = 120

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
astrodown = 'astrodown.commands:app'